+++ src/rpc/methods/net.rs
@@ -1,20 +1,19 @@
 // Copyright 2019-2024 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
+mod types;
+pub use types::*;
+
 use std::any::Any;
 use std::str::FromStr;
 
 use crate::libp2p::{NetRPCMethods, NetworkMessage, PeerId};
-use crate::lotus_json::lotus_json_with_self;
-use crate::rpc::{ApiPaths, Permission, ServerError};
-use crate::rpc::{Ctx, RpcMethod};
-use anyhow::Result;
+use crate::rpc::{ApiPaths, Ctx, Permission, RpcMethod, ServerError};
+use crate::utils::p2p::MultiaddrExt as _;
+use anyhow::{Context as _, Result};
 use cid::multibase;
 use futures::channel::oneshot;
 use fvm_ipld_blockstore::Blockstore;
-use libp2p::Multiaddr;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
 
 pub enum NetAddrsListen {}
 impl RpcMethod<0> for NetAddrsListen {
@@ -65,11 +64,47 @@ impl RpcMethod<0> for NetPeers {
             .into_iter()
             .map(|(id, addrs)| AddrInfo {
                 id: id.to_string(),
-                addrs,
+                addrs: addrs.into_iter().map(|addr| addr.without_p2p()).collect(),
             })
             .collect();
 
         Ok(connections)
+    }
+}
+
+pub enum NetFindPeer {}
+impl RpcMethod<1> for NetFindPeer {
+    const NAME: &'static str = "Filecoin.NetFindPeer";
+    const PARAM_NAMES: [&'static str; 1] = ["peer_id"];
+    const API_PATHS: ApiPaths = ApiPaths::V0;
+    const PERMISSION: Permission = Permission::Read;
+
+    type Params = (String,);
+    type Ok = AddrInfo;
+
+    // This is a no-op due to the fact that `rust-libp2p` implementation is very different to that
+    // in go. However it would be nice to investigate connection limiting options in Rust.
+    // See: <https://github.com/ChainSafe/forest/issues/4355>.
+    async fn handle(
+        ctx: Ctx<impl Blockstore>,
+        (peer_id,): Self::Params,
+    ) -> Result<Self::Ok, ServerError> {
+        let peer_id = PeerId::from_str(&peer_id)?;
+        let (tx, rx) = oneshot::channel();
+        let req = NetworkMessage::JSONRPCRequest {
+            method: NetRPCMethods::Peer(tx, peer_id),
+        };
+        ctx.network_send.send_async(req).await?;
+        let addrs = rx
+            .await?
+            .with_context(|| format!("peer {peer_id} not found"))?
+            .into_iter()
+            .map(|addr| addr.without_p2p())
+            .collect();
+        Ok(AddrInfo {
+            id: peer_id.to_string(),
+            addrs,
+        })
     }
 }
 
@@ -255,80 +290,5 @@ impl RpcMethod<1> for NetProtectAdd {
     ) -> Result<Self::Ok, ServerError> {
         let _ = PeerId::from_str(&peer_id)?;
         Ok(())
-    }
-}
-
-// Net API
-#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema)]
-#[serde(rename_all = "PascalCase")]
-pub struct AddrInfo {
-    #[serde(rename = "ID")]
-    pub id: String,
-    #[schemars(with = "ahash::HashSet<String>")]
-    pub addrs: ahash::HashSet<Multiaddr>,
-}
-
-lotus_json_with_self!(AddrInfo);
-
-#[derive(Debug, Default, Serialize, Deserialize, Clone, JsonSchema)]
-pub struct NetInfoResult {
-    pub num_peers: usize,
-    pub num_connections: u32,
-    pub num_pending: u32,
-    pub num_pending_incoming: u32,
-    pub num_pending_outgoing: u32,
-    pub num_established: u32,
-}
-lotus_json_with_self!(NetInfoResult);
-
-impl From<libp2p::swarm::NetworkInfo> for NetInfoResult {
-    fn from(i: libp2p::swarm::NetworkInfo) -> Self {
-        let counters = i.connection_counters();
-        Self {
-            num_peers: i.num_peers(),
-            num_connections: counters.num_connections(),
-            num_pending: counters.num_pending(),
-            num_pending_incoming: counters.num_pending_incoming(),
-            num_pending_outgoing: counters.num_pending_outgoing(),
-            num_established: counters.num_established(),
-        }
-    }
-}
-
-#[derive(Debug, Default, Serialize, Deserialize, Clone, JsonSchema)]
-#[serde(rename_all = "PascalCase")]
-pub struct NatStatusResult {
-    pub reachability: i32,
-    pub public_addrs: Option<Vec<String>>,
-}
-lotus_json_with_self!(NatStatusResult);
-
-impl NatStatusResult {
-    // See <https://github.com/libp2p/go-libp2p/blob/164adb40fef9c19774eb5fe6d92afb95c67ba83c/core/network/network.go#L93>
-    pub fn reachability_as_str(&self) -> &'static str {
-        match self.reachability {
-            0 => "Unknown",
-            1 => "Public",
-            2 => "Private",
-            _ => "(unrecognized)",
-        }
-    }
-}
-
-impl From<libp2p::autonat::NatStatus> for NatStatusResult {
-    fn from(nat: libp2p::autonat::NatStatus) -> Self {
-        use libp2p::autonat::NatStatus;
-
-        // See <https://github.com/libp2p/go-libp2p/blob/91e1025f04519a5560361b09dfccd4b5239e36e6/core/network/network.go#L77>
-        let (reachability, public_addrs) = match &nat {
-            NatStatus::Unknown => (0, None),
-            NatStatus::Public(addr) => (1, Some(vec![addr.to_string()])),
-            NatStatus::Private => (2, None),
-        };
-
-        NatStatusResult {
-            reachability,
-            public_addrs,
-        }
     }
 }
